requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.1
dj-database-url==2.3.0
Django==4.2.18
django-autocomplete-light==3.11.0
django-bootstrap-v5==1.0.11
django-smart-selects==1.7.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
